+++ src/jsontree.js
@@ -214,7 +214,7 @@
             var propertyName = properties[ propertyIndex ];
 
             if ( data.hasOwnProperty( propertyName ) ) {
-                renderValue( objectTypeContents, bindingOptions, propertyName, data[ propertyName ] );
+                renderValue( objectTypeContents, bindingOptions, propertyName, data[ propertyName ], propertyIndex === propertiesLength - 1 );
                 propertyCount++;
             }
         }
@@ -230,13 +230,13 @@
         for ( var dataIndex = 0; dataIndex < dataLength; dataIndex++ ) {
             var name = bindingOptions.useZeroIndexingForArrays ? dataIndex.toString() : ( dataIndex + 1 ).toString();
 
-            renderValue( objectTypeContents, bindingOptions, name, data[ dataIndex ] );
+            renderValue( objectTypeContents, bindingOptions, name, data[ dataIndex ], dataIndex === dataLength - 1 );
         }
 
         addArrowEvent( bindingOptions, arrow, objectTypeContents );
     }
 
-    function renderValue( container, bindingOptions, name, value ) {
+    function renderValue( container, bindingOptions, name, value, isLastItem ) {
         var objectTypeValue = createElement( container, "div", "object-type-value" ),
             arrow = bindingOptions.showArrowToggles ? createElement( objectTypeValue, "div", "no-arrow" ) : null,
             valueElement = null;
@@ -247,21 +247,32 @@
         if ( !isDefined( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "null", "null" );
 
+            createComma( bindingOptions, objectTypeValue, isLastItem );
+
         } else if ( isDefinedFunction( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "function", getFunctionName( value ) );
+            
+            createComma( bindingOptions, objectTypeValue, isLastItem );
 
         }  else if ( isDefinedBoolean( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "boolean", value );
+            
+            createComma( bindingOptions, objectTypeValue, isLastItem );
 
         } else if ( isDefinedNumber( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "number", value );
+            
+            createComma( bindingOptions, objectTypeValue, isLastItem );
 
         } else if ( isDefinedString( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "string", bindingOptions.showStringQuotes ? "\"" + value + "\"" : value );
+            
+            createComma( bindingOptions, objectTypeValue, isLastItem );
 
         } else if ( isDefinedDate( value ) ) {
             valueElement = createElementWithHTML( objectTypeValue, "span", "date", getCustomFormattedDateTimeText( value, bindingOptions.dateTimeFormat ) );
-            
+            createComma( bindingOptions, objectTypeValue, isLastItem );
+
         } else if ( isDefinedObject( value ) && !isDefinedArray( value ) ) {
             var objectTitle = createElement( objectTypeValue, "span", "object" ),
                 objectTypeContents = createElement( objectTypeValue, "div", "object-type-contents" ),
@@ -273,6 +284,8 @@
                 createElementWithHTML( objectTitle, "span", "count", "{" + propertyCount + "}" );
             }
 
+            createComma( bindingOptions, objectTitle, isLastItem );
+
         } else if ( isDefinedArray( value ) ) {
             var arrayTitle = createElement( objectTypeValue, "span", "array" ),
                 arrayTypeContents = createElement( objectTypeValue, "div", "object-type-contents" );
@@ -283,6 +296,7 @@
                 createElementWithHTML( arrayTitle, "span", "count", "[" + value.length + "]" );
             }
 
+            createComma( bindingOptions, arrayTitle, isLastItem );
             renderArrayValues( arrow, arrayTypeContents, bindingOptions, value );
         }
 
@@ -339,6 +353,12 @@
         return result;
     }
 
+    function createComma( bindingOptions, objectTypeValue, isLastItem ) {
+        if ( bindingOptions.showCommas && !isLastItem ) {
+            createElementWithHTML( objectTypeValue, "span", "comma", "," );
+        }
+    }
+
 
     /*
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -359,6 +379,7 @@
         options.showAllAsClosed = getDefaultBoolean( options.showAllAsClosed, false );
         options.sortPropertyNames = getDefaultBoolean( options.sortPropertyNames, true );
         options.sortPropertyNamesInAlphabeticalOrder = getDefaultBoolean( options.sortPropertyNamesInAlphabeticalOrder, true );
+        options.showCommas = getDefaultBoolean( options.showCommas, false );
 
         options = buildAttributeOptionStrings( options );
         options = buildAttributeOptionCustomTriggers( options );